morning_briefing/requirements.txt
# 晨间简报Agent - 依赖包

# HTTP请求
requests>=2.31.0
aiohttp>=3.9.0

# 数据处理
pandas>=2.1.0
numpy>=1.24.0

# API客户端
anthropic>=0.18.0
coingecko>=0.5.0

# RSS解析
feedparser>=6.0.10

# Discord通知
discord.py>=2.3.0

# Telegram通知
python-telegram-bot>=20.0

# sqlite3是Python内置模块，不需要安装

# 工具
python-dotenv>=1.0.0
pytz>=2023.3

# asyncio是Python内置模块，不需要安装

# 网页抓取
beautifulsoup4>=4.12.0

# JSON处理 (可选，用于更快解析)
# orjson>=3.9.0

# 时间处理
python-dateutil>=2.8.0
